// src/interfaces.ts
export interface ConstructorOptions {
  shelleyGenesisPath?: string;
  socketPath?: string;
  cliPath?: string;
  dir?: string;
  era?: string;
  network?: Network;
  httpProvider?: string;
  testnetMagic?: TestnetMagic;
}

export interface ProtocolParams {
  txFeePerByte: number;
  minUTxOValue: number;
  stakePoolDeposit: number;
  decentralization: number;
  poolRetireMaxEpoch: number;
  extraPraosEntropy: any;
  stakePoolTargetNum: number;
  maxBlockBodySize: number;
  maxTxSize: number;
  treasuryCut: number;
  minPoolCost: number;
  maxBlockHeaderSize: number;
  protocolVersion: {
    minor: number;
    major: number;
  };
  txFeeFixed: number;
  stakeAddressDeposit: number;
  monetaryExpansion: number;
  poolPledgeInfluence: number;
}

export interface AddressBuildOptions {
  paymentVkey?: string;
  stakeVkey?: string;
  paymentScript?: string;
  stakeScript?: string;
}

export interface StakePoolRegistrationOptions {
  pledge: number;
  margin: number;
  cost: number;
  url: string;
  metaHash: string;
  rewardAccount: string;
  owners: string[];
  relays: any[];
}

export interface CalculateMinFeeOptions {
  txBody: string;
  txIn: TxIn[];
  txOut: TxOut[];
  witnessCount: number;
}

export interface TransationSignOptions {
  signingKeys: string[];
  txBody: string;
}

export interface TransactionWitnessOptions {
  txBody: string;
  signingKey: string;
  scriptFile?: string;
}

export interface TransactionAssembleOptions {
  txBody: string;
  witnessFiles: string[];
}

export interface TransactionViewOptions {
  txBody?: string;
  txFile?: string;
}

export interface Transaction {
  txIn: TxIn[];
  txOut: TxOut[];
  txInCollateral?: TxInCollateral[];
  withdrawals?: Withdrawal[];
  certs?: Certificate[];
  fee?: number;
  mint?: Mint[];
  auxScript?: any;
  metadata?: any;
  invalidBefore?: number;
  invalidAfter?: number;
  scriptInvalid?: boolean;
  changeAddress?: string;
  witnessOverride?: string;
}

export class Metadata {
  [key: number]: {
    int?: number;
    string?: string;
    bytes?: string;
    list?: {
      int?: number;
      string?: string;
    }[];
    map?: {
      k: Metadata;
      v: Metadata;
    }[];
  };
}

export interface Tip {
  epoch: number;
  hash: string;
  slot: number;
  block: number;
  era: string;
  syncProgress: string;
}
// ...
export interface AddressInfo {
  address: string;
  era: string;
  encoding: string;
  type: string;
  base16: string;
}

export interface TxIn extends TxInCollateral {
  script?: JSONValue;
  datum?: JSONValue;
  redeemer?: JSONValue;
  executionUnits?: string[];
}

export interface TxOut {
  address: string;
  value: { [key: string]: string };
  datumHash: string;
}

export interface TxInCollateral {
  txHash: string;
  txId: string;
}

export interface Withdrawal {
  stakingAddress: string;
  reward: number;
  script: JSONValue;
  datum: JSONValue;
  redeemer: JSONValue;
  executionUnits: string[];
}

export interface Certificate {
  cert: string;
  script: JSONValue;
  datum: JSONValue;
  redeemer: JSONValue;
  executionUnits: string[];
}

export interface Mint {
  action: string;
  quantity: string;
  asset: string;
  script: JSONValue;
  datum: JSONValue;
  redeemer: JSONValue;
  executionUnits: string[];
}

export interface Utxo {
  txHash: string;
  txId: number;
  value: { [key: string]: string };
  datumHash?: string;
}